created separate file for logger
