+++ kutana/__init__.py
@@ -1,14 +1,14 @@
+from kutana.logger import logger
+
 from kutana.kutana import *
 from kutana.executor import *
 from kutana.plugins.plugin import *
-
-from kutana.tools.functions import *
 
 from kutana.controllers.vk import *
 from kutana.controllers.dumping import *
 
 from kutana.environments.vk import *
-
+from kutana.tools.functions import *
 from kutana.shortcuts import *
 
 __version__ = "0.0.1"
